Bug #6253: Add canvas image extraction prompt.
This patch was later revised for Bug #12684 to change the default selected
notification choice to "Not Now", to add a new notification icon, and to
improve message text and message formatting.

The notification image is freely licensed and obtainable from:
https://openclipart.org/image/300px/svg_to_png/21620/ben_palette.png

NOTE: Applying this patch requires an additional patch to TorButton, to
store the additional UI strings before localisation.

This patch was later revised for Bug #7265 to log the script urls that request
canvas access to the Browser Console.

The notification message is only displayed for first parties that try to
access the canvas. Third party access is simply blocked and logged, without
permission notification.

The Bug #12684 changes were written by Isis Lovecruft.
The Bug #7265 changes were written by Mike Perry.