Add security enhancements suggested by Jesse Ruderman.
We were going to put these in the security slider, but perhaps they should be
set by default. There will be a perf impact, but let's see if anyone notices
and complains, at least for a beta or two. If not, maybe we can even set these
in a release.