Bug 13049: updater fails (self.update is undefined).
Replace "self.update" with "this.update" in two add-on compatibility check
functions.  Using self.update (which is undefined) caused the updater to
silently fail if one or more installed add-ons were marked in their
install.rdf file as not compatible with the new platformVersion.

We also replaced "TBB" with "Tor Browser" in the about box text.