Bug 12811: Add MOZ_D3DCOMPILER_DLL iff available.
It seems MOZ_D3DCOMPILER_DLL is not avilable when compiling with
mingw-w64. Trying to include it nevertheless breaks the packaging step
as `@BINPATH@/@MOZ_D3DCOMPILER_DLL@` resolves to `bin/` which results
in including all the things beginning with a `bin/` (again) which is
treated as an error in case they got already included.

This got fixed more or less by accident by bug 999260 which did not make
it into ESR 31 (https://hg.mozilla.org/mozilla-central/rev/8b48386fc226).