Bug 12827: Create preference to disable SVG.
If the svg.in-content.enabled preference is false, disallow all use of
SVG within content pages.

In the following situations it is very difficult to determine if code
is executing within a chrome context or not:
  SVG hasFeature() API.
  SVG hasExtension() API.
  Use of SVG glyphs within custom OpenType fonts.
In these cases, everything is assumed to be content; that is, setting
the pref. to false will block use of the above features from chrome
as well. This is OK because these features are unlikely to be used by
core browser code.

Also disallow use of SVG in favicons (bug 18602).

The browser does some negative caching of failed favicon loads, so
SVG favicons may not begin to work on all pages immediately after
svg.in-content.enabled is changed from false to true.