fixup! Bug 13252 - Do not store data in the app bundle
On Mac OS, store user data under
~/Library/Application Support/TorBrowser-Data if the application
is installed in a directory whose path contains "/Applications" or
if the user lacks write permission to the directory that contains
TorBrowser.app.

Since the TorBrowser-Data directory may be shared among different
installations of the application on Mac OS, embed the app path in
the update dir so that the update history is partitioned.