squash! Bug 6253: Add canvas image extraction prompt.
When processing the ShowCanvasPermissionPrompt message in the main
(parent) process, avoid returning an error from the IPC handler if
the browser element cannot be obtained. Prior to this change, canvas
prompt requests that arrived as a tab was closing would generate
an error, which in turn caused the main process to kill the content
process.