Bug 23016: "Print to File" does not create the expected file in non-English locales
The Problem:

During firefox and Web Content process startup,
::OverrideDefaultLocaleIfNeeded() is called which will conditionally
::setlocale() to "C.UTF-8" or "C" based off of the
javascript.use_us_english_locale preference (to prevent fingerprinting
based on how dates and what-not are formatted).  Sometime after this
call in the Web Content process the locale is set to the system's locale
which effectively stomps over the work done by the
::OverrideDefaultLocaleIfNeeded() call.  As a result, the firefox
process and the Web Content process are configured to use different
locales; firefox uses "C.UTF-8" while Web Content uses "" (system
default).

On Linux, the "Print to File" printer is a default GTK printer whose
name is localized based off of the process's locale.  The GTK print dialog
is hosted in the firefox process and therefore the printer name will be
'Print to File.'  This process sends this name over to the Web Content
process, which iterates over the list of printers and finds one whose
name matches.  However, as the Web Content process's locale is set to
system, its printer names will be localized and in the language of the
system, so no printer with the name 'Print to File' will be found, and
printing fails.

This is why disabling javascript.use_us_english_locale fixes the issue.
It also explains why disabling multi-process mode via the
browser.tabs.remote.autostart.2 preference fixes the issue; one
process means the printer-name is never used as a selector and also
means no mismatched locale can happen.

The Solution:

The fix for this issue is to first check the
javascript.use_us_english_locale preference each place in code where
setlocale occurs, particularly in the nsLocaleService object which is the
particular code path which overrides setlocale in the Web Content
process.