Bug 25603 - Upgrade Orfox's https-everywhere add-on
We bundle https-everywhere in-tree and add it using Fennec's
distribution system. The newer webextensions versions of
https-everywhere do not work with this version of Fennec, so we
are using https-everywhere's last extension release (v5.2.21).

Unfortunately, this version was released 18 July, 2017 so the ruleset
is old. As a workaround, we compiled v5.2.21 and included the ruleset
from the 2018.4.11 tag [0]. This created a byte-for-byte reproducible
XPI [1].

[0] https://trac.torproject.org/projects/tor/ticket/25603#comment:10
[1] 7a33c13dbd80fd881b1508fca6dc10fca787f8eb4da754104321537240ffb866